added alerts for deletes

+++ src/components/ToolEdit/ToolEdit.js
@@ -1,6 +1,7 @@
 import React, { Component } from 'react';
 import axios from 'axios';
 import Dropzone from 'react-dropzone';
+import swal from 'sweetalert2';
 import defaultTool from './../../assets/defaultTool.png';
 
 export default class ToolEdit extends Component {
@@ -193,9 +194,44 @@ export default class ToolEdit extends Component {
   }
 
   delete() {
-    axios.delete(`/api/delete/tool/${this.props.match.params.id}`)
-    axios.delete(`/api/tooltags/${this.props.match.params.id}`)
-    this.props.history.push('/search');
+    const swalWithBootstrapButtons = swal.mixin({
+      confirmButtonClass: 'btn btn-success',
+      cancelButtonClass: 'btn btn-danger',
+      buttonsStyling: false,
+
+    })
+
+    swalWithBootstrapButtons({
+      title: 'Are you sure?',
+      text: "You won't be able to revert this!",
+      type: 'warning',
+      background: '#252525',
+      showCancelButton: true,
+      confirmButtonText: 'Yes',
+      cancelButtonText: 'No',
+      reverseButtons: true
+    }).then((result) => {
+      if (result.value) {
+        axios.delete(`/api/delete/tool/${this.props.match.params.id}`)
+        axios.delete(`/api/tooltags/${this.props.match.params.id}`)
+        this.props.history.push('/search');
+        swalWithBootstrapButtons({
+          title: 'Deleted!',
+          text: 'Your tool has been deleted.',
+          type: 'success',
+          background: '#252525'
+        })
+      } else if (
+        result.dismiss === swal.DismissReason.cancel
+      ) {
+        swalWithBootstrapButtons({
+          title: 'Cancelled',
+          text: 'Your tool is safe.',
+          type: 'error',
+          background: '#252525'
+        })
+      }
+    })
   }
 
   render() {
